# Make `ForEach-Object` faster for its commonly used scenarios