@daxian-dbw 2019-08-29
> Is it simply not possible to perform optimizations such as this one transparently to the user and the debugger by doing the work in the AST itself?

The original AST should be preserved exactly as is because it should reflect exactly what the user typed on the console or wrote in the file.
If your thought is to generate another AST via a visitor before compiling, and use that one for compilation, then there is no guarantee the extent will be exactly the same and it will be transparent to the debugger.
Take this change as an example, once `foreach-object { ... }` is replaced by `. { process { ... } }` in the generated AST, you will have problem on the extent of the `process {` part, because that's new elements added the generated AST. Also, there won't be a `CommandProcessor` representing the `ForEach-Object` in the pipeline, and thus the debugger will break at the open curly instead of `ForEach-Object` for `Set-PSBreakpoint -Command Foreach-Object`.

For this specific change, I think there is something we can try if we really want to make it work with debugger: when rewriting the pipeline, instead of creating a command processor with the script block, we can construct a `FunctionInfo` with the name `ForEach-Object` and make its extent the same as the `ForEach-Object` extent. That might work, and probably worth explore as a follow-up PR.