@rjmholt 2019-08-29
@KirkMunro 

> Is it simply not possible to perform optimizations such as this one transparently to the user and the debugger by doing the work in the AST itself? Or is it possible, but an approach like that is considered undesirable for other reasons

It's definitely possible. There's a good example in `DynamicKeywordStatementAst`:

https://github.com/PowerShell/PowerShell/blob/704ee90d5cd2909ebc98a8d624359eed5d1b1cbc/src/System.Management.Automation/engine/parser/ast.cs#L6817-L7027

But to my mind, the questions are those of layering:

- How faithfully should the AST represent the program as it was given to the parser?
- How much should the AST know about what consumes it?
- To what extent (heh) should I be able to reconstruct a program from its AST?
- If we add syntactic sugar, does the parser own it, the AST or the compiler?

These problems (/restatements of the same problem) are inconsistently addressed in PowerShell:

- [The parser doesn't record where traps occur in a block at all, so the AST is created with some arguably important information missing](https://github.com/PowerShell/PowerShell/blob/704ee90d5cd2909ebc98a8d624359eed5d1b1cbc/src/System.Management.Automation/engine/parser/Parser.cs#L1858-L1908)
- [Dynamic keyword invocations get their own AST, but they do their own AST transformation to be compiled](https://github.com/PowerShell/PowerShell/blob/704ee90d5cd2909ebc98a8d624359eed5d1b1cbc/src/System.Management.Automation/engine/parser/Compiler.cs#L2978-L2986)
- The compiler's `*Ops` libraries have some syntactic sugar transformations in them, like [for job control operators (`Get-Command &`)](https://github.com/PowerShell/PowerShell/blob/704ee90d5cd2909ebc98a8d624359eed5d1b1cbc/src/System.Management.Automation/engine/parser/Compiler.cs#L3092-L3098) and here with the `ForEach-Object` now

If we were writing it all over today, we might have a syntactic sugar phase/layer between the parser and the compiler implemented as an AstVisitor that takes in one AST and spits out the transformed one, although the immutability of the AST (which I still think is a very good thing) makes that a bit more expensive. That's certainly the way other statically compiled. languages have gone.

But I do feel ambivalent about syntactic sugar (in the strict sense of a language-level AST-to-AST transformation, where one syntax is a macro for another) in an interpreted language. I'm not sure about Python, Ruby and friends, but both C# and Java seem to have a policy against it as far as I can tell; they go direct to bytecode and leave the optimisations to the runtime.

For PowerShell though, I think (1) we want to encourage good practice with syntax, (2) it's good to try and find performance wherever we can while still finding time for other work.

So basically, I don't think we have a good, consistent answer for how something like this should be implemented, but this implementation is one of the better ones in terms of both how it's done and where it sits.