@iSazonov 2019-08-30
I like the optimization because it works well for _all_ scenarios.
Although it’s obvious that there could be optimizations that can work well in a particular situation. That's why the question of Ast optimizations arises.

For example, for a simple management script in a task scheduler, an optimization is not to do any optimization at all since the interpreter will be the fastest.
For data manning script (parse and analyze large log file(s)), we could say in advance `pwsh -Optimization Full` to apply all optimizations _before_ run. In the scenario generating optimized Ast tree could bring significant acceleration. Currently single optimization we do for such scenarios - switch from interpreter to compiler for long running script blocks.

> they go direct to bytecode and leave the optimizations to the runtime.

Since we have new modern runtime we could review if we can generate code so that it can optimize by the runtime better.