@KirkMunro 2019-09-03
@daxian-dbw This PR introduces a breaking change. :boom:

Step 1: Create a file with the following contents:

```PowerShell
1..2 | ForEach-Object{

    Write-Host('MyInvocation in ForEach-Object: ' + $MyInvocation.MyCommand.Name)

}
```

Step 2: Dot-source or invoke that file with the call operator.

Before this change (using preview 3):
`$MyInvocation.MyCommand.Name` results in the name of the file.

After this change (using a custom build with the latest bits, pulled just now):
`$MyInvocation.MyCommand.Name` is `$null`.