@daxian-dbw 2019-09-03
Interesting, I knew there might be cases that I didn't catch :) I will be working on a fix.

I guess we also need an answer to this question:
> why does the behaviour differ between Where-Object and ForEach-Object? They are both cmdlets, and I don't think it's unreasonable for users to expect a consistent experience between these two commands in terms of how their scriptblock implementations behave.