@daxian-dbw 2019-09-05
#10485 was submitted to revert this PR given that it introduced a breaking change to the value of `$MyInvocation` within the script block argument.